# api/openapi.yaml
openapi: 3.0.3
info:
  title: AgentFlow Runtime API
  description: |
    OpenAPI contract for endpoints currently exposed by `cmd/agentflow` runtime.

    Notes:
    - API key authentication is accepted via `X-API-Key` header.
    - Query-string `api_key` is disabled by default for security reasons.
  version: 1.3.0
  contact:
    name: AgentFlow Team
    url: https://github.com/BaSui01/agentflow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: HTTP API server
...
paths:
...
  /api/v1/chat/completions:
    post:
      tags: [Chat]
      summary: Chat completion
      description: |
        Send a chat completion request to the configured LLM provider.
        Note: This endpoint is only available when an LLM API key is configured.
      x-conditional: "Requires LLM API key configuration (config.llm.api_key)"
      operationId: chatCompletion
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Chat completion response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '400':
          description: Invalid request
        '500':
          description: Internal error
  /api/v1/chat/completions/stream:
    post:
      tags: [Chat]
      summary: Streaming chat completion (SSE)
      description: |
        Send a streaming chat completion request via Server-Sent Events.
        Note: This endpoint is only available when an LLM API key is configured.
      x-conditional: "Requires LLM API key configuration (config.llm.api_key)"
      operationId: chatCompletionStream
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: SSE stream of chat completion chunks
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          description: Invalid request
        '500':
          description: Internal error
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Envelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          additionalProperties: true
        error:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          description: Unique request identifier for tracing
          example: "req-abc123"
...
    ChatRequest:
      type: object
      required: [model, messages]
      properties:
        trace_id:
          type: string
          description: Trace ID for request tracing
          example: "trace-123"
        tenant_id:
          type: string
          description: Tenant ID for multi-tenancy
          example: "tenant-1"
        user_id:
          type: string
          description: User identity
          example: "user-1"
        model:
          type: string
          description: Model name (e.g. gpt-4, claude-3-opus)
          example: "gpt-4"
        messages:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [system, user, assistant, tool]
              content:
                type: string
        max_tokens:
          type: integer
          description: Maximum number of tokens to generate
          example: 4096
        temperature:
          type: number
          description: Sampling temperature (0-2)
          example: 0.7
        top_p:
          type: number
          description: Nucleus sampling parameter (0-1)
          example: 1.0
        stop:
          type: array
          items:
            type: string
          description: Stop sequences
        tools:
          type: array
          items:
            $ref: '#/components/schemas/ToolSchema'
          description: Available tools for function calling
        tool_choice:
          type: string
          description: Tool selection mode (auto, none, or specific tool name)
          example: "auto"
        timeout:
          type: string
          description: Request timeout duration
          example: "30s"
        metadata:
          type: object
          additionalProperties:
            type: string
          description: Custom metadata key-value pairs
        tags:
          type: array
          items:
            type: string
          description: Routing tags
...
    ToolSchema:
      type: object
      required: [name, parameters]
      properties:
        name:
          type: string
          description: Tool name
          example: "get_weather"
        description:
          type: string
          description: Tool description
          example: "Get the current weather for a location"
        parameters:
          type: object
          description: JSON Schema for tool parameters
          additionalProperties: true
        version:
          type: string
          description: Tool version
          example: "1.0.0"